+++ CHANGELOG.md
@@ -13,4 +13,5 @@ All notable changes to this project will be documented in this file.
 - Simple example `helloworld` project under `examples/` ([#62](https://github.com/google/jib/pull/62))
 
 ### Fixed
+- Infers common credential helper names (for GCR and ECR) ([#64](https://github.com/google/jib/pull/64))
 - Building applications with no resources ([#73](https://github.com/google/jib/pull/73))